refactor: 로또 설정 enum 디렉토리 이동
