+++ src/main/java/lotto/application/RandomNumberGenerator.java
@@ -1,8 +1,8 @@
 package lotto.application;
 
-import static lotto.domain.LottoConfig.LOTTO_COUNT;
-import static lotto.domain.LottoConfig.MAX_LOTTO_NUMBER;
-import static lotto.domain.LottoConfig.MIN_LOTTO_NUMBER;
+import static lotto.enums.LottoConfig.LOTTO_COUNT;
+import static lotto.enums.LottoConfig.MAX_LOTTO_NUMBER;
+import static lotto.enums.LottoConfig.MIN_LOTTO_NUMBER;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
