+++ src/test/java/lotto/domain/LottoNumberTest.java
@@ -1,8 +1,8 @@
 package lotto.domain;
 
 import static lotto.enums.ErrorMassage.OUT_OF_RANGE_LOTTO_NUMBER;
-import static lotto.domain.LottoConfig.MAX_LOTTO_NUMBER;
-import static lotto.domain.LottoConfig.MIN_LOTTO_NUMBER;
+import static lotto.enums.LottoConfig.MAX_LOTTO_NUMBER;
+import static lotto.enums.LottoConfig.MIN_LOTTO_NUMBER;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
